@@ -18,6 +18,7 @@ class RepaymentSchedule extends Model
     ];
 
     public function loan() {
-        return $this->belongsTo('App\Models\Loan');
+        return $this->belongsTo(Loan::class);
     }
+
 }
