@@ -2,6 +2,13 @@
 
 namespace App\Providers;
 
+use App\Events\LoanApplicationApproved;
+use App\Events\NewLoanApplication;
+use App\Events\RepaymentSuccessful;
+use App\Listeners\CheckLoanRepaidAndMailCustomer;
+use App\Listeners\SendMailAdminNewLoanApplication;
+use App\Listeners\SendMailCustomerLoanApproved;
+use App\Listeners\SendPaymentSuccessfulMailCustomer;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
@@ -10,7 +17,7 @@ use Illuminate\Support\Facades\Event;
 class EventServiceProvider extends ServiceProvider
 {
     /**
-     * The event listener mappings for the application.
+     * The event to listener mappings for the application.
      *
      * @var array<class-string, array<int, class-string>>
      */
@@ -18,24 +25,30 @@ class EventServiceProvider extends ServiceProvider
         Registered::class => [
             SendEmailVerificationNotification::class,
         ],
+        NewLoanApplication::class => [
+            SendMailAdminNewLoanApplication::class,
+        ],
+        LoanApplicationApproved::class => [
+            SendMailCustomerLoanApproved::class
+        ],
+        RepaymentSuccessful::class => [
+            SendPaymentSuccessfulMailCustomer::class,
+            CheckLoanRepaidAndMailCustomer::class
+        ],
     ];
 
     /**
      * Register any events for your application.
-     *
-     * @return void
      */
-    public function boot()
+    public function boot(): void
     {
         //
     }
 
     /**
      * Determine if events and listeners should be automatically discovered.
-     *
-     * @return bool
      */
-    public function shouldDiscoverEvents()
+    public function shouldDiscoverEvents(): bool
     {
         return false;
     }
