@@ -2,7 +2,10 @@
 
 namespace Tests\Feature\Auth;
 
+use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use Tests\TestCase;
 
 class LogoutTest extends TestCase
@@ -14,18 +17,18 @@ class LogoutTest extends TestCase
         $user = $this->createUser();
         $this->authUser($user);
 
-        $response = $this->getJson(route('user.logout'))
+        $response = $this->postJson(route('customer.logout'))
                     ->assertOk()
                     ->json();
 
-        $this->assertEquals('SUCCESS', $response['status']);
+        $this->assertEquals('success', $response['status']);
     }
 
     public function test_user_try_to_logout_without_login()
     {
         $this->withExceptionHandling();
 
-        $response = $this->getJson(route('user.logout'))
+        $response = $this->postJson(route('customer.logout'))
                     ->assertUnauthorized();
     }
 }
