@@ -2,6 +2,8 @@
 
 namespace Tests\Feature;
 
+use App\Models\Loan;
+use App\Models\RepaymentSchedule;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
@@ -34,32 +36,36 @@ class LoanRepaymentTest extends TestCase
         $this->_userLoan = $this->createLoan(['user_id' => $this->_user->id]);
 
         $this->authUser($this->_adminUser);
-        $this->patchJson(route('loan.approve', ['loan_id' => $this->_userLoan->id]));
+        $this->postJson(route('loan.approve', ['loan_id' => $this->_userLoan->id]));
     }
     
     public function test_user_repay_first_term_amount()
     {
         $this->authUser($this->_user);
+        $amount = sprintf('%0.2f', $this->_userLoan->amount / $this->_userLoan->loan_term);
 
-        $response = $this->postJson(route('loan.repayment'), ['amount' => $this->_userLoan->amount / $this->_userLoan->loan_term, 'loan_id' => $this->_userLoan->id])
+        $this->postJson(route('loan.repayment'), ['amount' => $amount, 'loan_id' => $this->_userLoan->id])
                     ->assertOk()
                     ->json();
 
-        $this->assertDatabaseHas('repayment_schedules', ['loan_id' => $this->_userLoan->id, 'status' => 'paid']);
+        $repayment = RepaymentSchedule::where('loan_id', $this->_userLoan->id)->first();
+
+        $this->assertEquals(sprintf('%0.2f', $repayment->repayment_amount), $amount);
+        $this->assertEquals('paid', $repayment->status);
     }
 
     public function test_user_repay_all_term_amount()
     {
         $this->authUser($this->_user);
 
-        $termAmount = $this->_userLoan->amount / $this->_userLoan->loan_term;
+        $termAmount = round($this->_userLoan->amount / $this->_userLoan->loan_term, 2);
 
         for($i = 0; $i < $this->_userLoan->loan_term; $i++) {
-            
-            $response = $this->postJson(route('loan.repayment'), ['amount' => $termAmount, 'loan_id' => $this->_userLoan->id])
+            $response = $this->postJson(route('loan.repayment'), ['amount' => sprintf('%0.2f', $termAmount), 'loan_id' => $this->_userLoan->id])
             ->assertOk()
             ->json();
         }
+        $loan = Loan::find($this->_userLoan->id)->first();
         
         $this->assertDatabaseHas('loans', ['id' => $this->_userLoan->id, 'status' => 'paid']);
     }
@@ -67,13 +73,14 @@ class LoanRepaymentTest extends TestCase
     public function test_user_repay_less_term_amount()
     {
         $this->authUser($this->_user);
+        $amount = round($this->_userLoan->amount / $this->_userLoan->loan_term, 2);
 
-        $response = $this->postJson(route('loan.repayment'), ['amount' => ($this->_userLoan->amount / $this->_userLoan->loan_term) - 2, 'loan_id' => $this->_userLoan->id])
+        $response = $this->postJson(route('loan.repayment'), ['amount' => sprintf('%0.2f', $amount - 2), 'loan_id' => $this->_userLoan->id])
                     ->assertOk()
                     ->json();
 
-        $this->assertEquals('FAIL', $response['status']);
-        $this->assertEquals('Repayment amount is insufficient.', $response['message']);
+        $this->assertEquals('failed', $response['status']);
+        $this->assertEquals('Loan amount is insufficient for repayment.', $response['message']);
 
     }
     
@@ -83,11 +90,12 @@ class LoanRepaymentTest extends TestCase
 
         $user = $this->createUser();
         $this->authUser($user);
+        $amount = round($this->_userLoan->amount / $this->_userLoan->loan_term, 2);
 
-        $response = $this->postJson(route('loan.repayment'), ['amount' => $this->_userLoan->amount / $this->_userLoan->loan_term, 'loan_id' => $this->_userLoan->id])
-                    ->assertUnauthorized()
+        $response = $this->postJson(route('loan.repayment'), ['amount' => sprintf('%0.2f', $amount), 'loan_id' => $this->_userLoan->id])
+                    ->assertNotFound()
                     ->json();
                 
-        $this->assertEquals('FAIL', $response['status']);
+        $this->assertEquals('failed', $response['status']);
     }
 }
